chore: fix tests

+++ test/commands/lwc/test/run.nut.ts
@@ -21,9 +21,11 @@ describe('lightning:lwc:test:run', () => {
 
   before('prepare session and ensure environment variables', async () => {
     testSession = await TestSession.create({
-      setupCommands: ['sfdx force:lightning:lwc:test:setup'],
       project: { gitClone: 'https://github.com/trailheadapps/dreamhouse-lwc' },
     });
+    if (!testSession.project) {
+      throw new Error(`Expected project: ${testSession.project}`);
+    }
     // I'm not sure why this test started failing
     // I was unable to reproduce it locally, even in the generated test_session directory
     fs.rmSync(
@@ -48,6 +50,9 @@ describe('lightning:lwc:test:run', () => {
     const output = execCmd<RunResult>('force:lightning:lwc:test:run --json', {
       ensureExitCode: 0,
     }).jsonOutput;
+    if (!output) {
+      throw new Error(`Expected output: ${output}`);
+    }
     expect(output.result.message).to.equal('Test run complete. Exited with status code: 0');
     expect(output.result.jestExitCode).to.equal(0);
   });
@@ -63,6 +68,9 @@ describe('lightning:lwc:test:run', () => {
   });
 
   it('properly displays failed tests (human)', async () => {
+    if (!testSession.project) {
+      throw new Error(`Expected project: ${testSession.project}`);
+    }
     const testPath = path.join(
       testSession.project.dir,
       'force-app',
